ui: modify top bar, bottom bar, search bar and filter button
This commit worked on trying to integrate available UI template. Most of
their code used xml, flutter or out-of-date compose material version,
which made me to fix a lot of stuff, up to the point that I kind of gave
up. And also finding a template that match to the app design is
difficult and tedious.

At the end of the day, I only modified the top bar, bottom bar, search
bar and filter button, and some change in coloring.